# optimize logsoftmax