--- a/spark/dl/src/main/scala/com/intel/analytics/bigdl/nn/LogSoftMax.scala
+++ b/spark/dl/src/main/scala/com/intel/analytics/bigdl/nn/LogSoftMax.scala
@@ -16,8 +16,11 @@
 
 package com.intel.analytics.bigdl.nn
 
+import java.util
+
+import com.intel.analytics.bigdl.mkl.MKL
 import com.intel.analytics.bigdl.nn.abstractnn.TensorModule
-import com.intel.analytics.bigdl.tensor.Tensor
+import com.intel.analytics.bigdl.tensor.{DoubleType, FloatType, Storage, Tensor}
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
 import com.intel.analytics.bigdl.utils.Engine
 
@@ -39,11 +42,16 @@ class LogSoftMax[T: ClassTag](
   implicit ev: TensorNumeric[T]) extends TensorModule[T] {
   @transient
   private var results: Array[Future[Unit]] = null
+  @transient
+  private var buffer1: Array[T] = null
+  @transient
+  private var buffer2: Array[T] = null
+
 
   override def updateOutput(input: Tensor[T]): Tensor[T] = {
     require(input.dim() == 1 || input.dim() == 2,
       "LogSoftMax: " + ErrorInfo.constrainInputAsVectorOrBatch)
-    output.resizeAs(input)
+    output.resizeAs(input).copy(input)
     val (nframe, dim) =
       if (input.nDimension() == 1) (1, input.size(1)) else (input.size(1), input.size(2))
 
@@ -68,63 +76,158 @@ class LogSoftMax[T: ClassTag](
   }
 
   private def updateOutputFrame(in: Tensor[T], out: Tensor[T]): Unit = {
-    var logsum = ev.fromType[Int](0)
-    val maxInput = in.max()
-    in.apply1(v => {
-      logsum = ev.plus(logsum, ev.exp(ev.minus(v, maxInput))); v
-    })
-    logsum = ev.plus(maxInput, ev.log(logsum))
+    if (buffer1 == null || buffer1.length < in.nElement) {
+      buffer1 = Array.fill(in.nElement)(ev.fromType[Int](1))
+    }
+    if (buffer2 == null || buffer2.length < in.nElement) {
+      buffer2 = new Array[T](in.nElement)
+    }
+    ev.getType match {
+      case FloatType =>
 
-    out.map(in, (outData, inData) => {
-      ev.minus(inData, logsum)
-    })
+        MKL.vsExp(in.nElement,
+          in.storage.asInstanceOf[Storage[Float]].array,
+          in.storageOffset - 1,
+          buffer2.asInstanceOf[Array[Float]],
+          0)
+
+        val dot = MKL.vsdot(in.nElement,
+          buffer2.asInstanceOf[Array[Float]],
+          0,
+          1,
+          buffer1.asInstanceOf[Array[Float]],
+          0,
+          1)
+        val sum = ev.negative(
+          ev.log(ev.fromType[Float](dot)))
+
+        MKL.vsaxpy(in.nElement,
+          ev.toType[Float](sum),
+          buffer1.asInstanceOf[Array[Float]],
+          0,
+          1,
+          out.storage.asInstanceOf[Storage[Float]].array,
+          out.storageOffset - 1,
+          1)
+
+      case DoubleType =>
+
+        MKL.vdExp(in.nElement,
+          in.storage.asInstanceOf[Storage[Double]].array,
+          in.storageOffset - 1,
+          buffer2.asInstanceOf[Array[Double]],
+          0)
+
+        val dot = MKL.vddot(in.nElement,
+          buffer2.asInstanceOf[Array[Double]],
+          0,
+          1,
+          buffer1.asInstanceOf[Array[Double]],
+          0,
+          1)
+        val sum = ev.negative(
+          ev.log(ev.fromType[Double](dot)))
+
+        MKL.vdaxpy(in.nElement,
+          ev.toType[Double](sum),
+          buffer1.asInstanceOf[Array[Double]],
+          0,
+          1,
+          out.storage.asInstanceOf[Storage[Double]].array,
+          out.storageOffset - 1,
+          1)
+      case _ => throw new UnsupportedOperationException(s"Only Float/Double supported")
+    }
+    out
   }
 
   override def updateGradInput(input: Tensor[T], gradOutput: Tensor[T]): Tensor[T] = {
     require(output.nDimension() == 1 || output.nDimension() == 2, "vector or matrix expected")
     require(gradOutput.dim() == input.dim(), "LogSoftMax: input and gradOutput shapes do not " +
       "match, input_dim: " + input.dim() + ", gradOutput_dim: " + gradOutput.dim())
-    gradInput.resizeAs(input)
+    gradInput.resizeAs(input).copy(gradOutput)
     val (nframe, dim) =
       if (output.nDimension() == 1) (1, output.size(1)) else (output.size(1), output.size(2))
 
-    if (results == null || results.length != nframe) {
-      results = new Array[Future[Unit]](nframe)
+
+    if (nframe == 1) {
+      updateGradInputFrame(output, gradInput)
+    } else {
+      if (results == null || results.length != nframe) {
+        results = new Array[Future[Unit]](nframe)
+      }
+      var t = 1
+      while (t <= nframe) {
+        val _t = t
+        results(_t - 1) = Engine.model.invoke(() => {
+          updateGradInputFrame(output.select(1, _t), gradInput.select(1, _t))
+        })
+        t += 1
+      }
+      Engine.model.sync(results)
     }
-
-    var t = 1
-    while (t <= nframe) {
-      val _t = t
-      results(_t - 1) = Engine.model.invoke(() => {
-        var sum = 0.0
-        var d = 1
-        while (d <= dim) {
-          if (gradOutput.dim() == 1) {
-            sum += ev.toType[Double](gradOutput.valueAt(d))
-          } else {
-            sum += ev.toType[Double](gradOutput.valueAt(_t, d))
-          }
-          d += 1
-        }
-
-        d = 1
-        while (d <= dim) {
-          if (input.dim() == 1) {
-            gradInput.setValue(d, ev.minus(gradOutput.valueAt(d),
-              ev.times(ev.exp(output.valueAt(d)), ev.fromType[Double](sum))))
-          } else {
-            gradInput.setValue(_t, d, ev.minus(gradOutput.valueAt(_t, d),
-              ev.times(ev.exp(output.valueAt(_t, d)), ev.fromType[Double](sum))))
-          }
-          d += 1
-        }
-      })
-      t += 1
-    }
-    Engine.model.sync(results)
-
     gradInput
   }
+
+  private def updateGradInputFrame(out: Tensor[T], gradOut: Tensor[T]): Unit = {
+    ev.getType match {
+       case FloatType =>
+
+         MKL.vsExp(out.nElement,
+           out.storage.asInstanceOf[Storage[Float]].array,
+           out.storageOffset - 1,
+           buffer2.asInstanceOf[Array[Float]],
+           0)
+
+         val dot = MKL.vsdot(gradOut.nElement,
+           gradOut.storage.asInstanceOf[Storage[Float]].array,
+           gradOut.storageOffset - 1,
+           1,
+           buffer1.asInstanceOf[Array[Float]],
+           0,
+           1)
+         val sum = ev.negative(
+           ev.fromType[Float](dot))
+
+         MKL.vsaxpy(gradOut.nElement,
+           ev.toType[Float](sum),
+           buffer2.asInstanceOf[Array[Float]],
+           0,
+           1,
+           gradOut.storage.asInstanceOf[Storage[Float]].array,
+           gradOut.storageOffset - 1,
+           1)
+
+      case DoubleType =>
+
+        MKL.vdExp(out.nElement,
+          out.storage.asInstanceOf[Storage[Double]].array,
+          out.storageOffset - 1,
+          buffer2.asInstanceOf[Array[Double]],
+          0)
+
+        val dot = MKL.vddot(gradOut.nElement,
+          gradOut.storage.asInstanceOf[Storage[Double]].array,
+          gradOut.storageOffset - 1,
+          1,
+          buffer1.asInstanceOf[Array[Double]],
+          0,
+          1)
+        val sum = ev.negative(
+          ev.fromType[Double](dot))
+
+        MKL.vdaxpy(gradOut.nElement,
+          ev.toType[Double](sum),
+          buffer2.asInstanceOf[Array[Double]],
+          0,
+          1,
+          gradOut.storage.asInstanceOf[Storage[Double]].array,
+          gradOut.storageOffset - 1,
+          1)
+      case _ => throw new UnsupportedOperationException(s"Only Float/Double supported")
+    }
+  }
+
 }
 
 object LogSoftMax {