--- a/spark/dl/src/test/scala/com/intel/analytics/bigdl/models/ModelGraientCheckSpec.scala
+++ b/spark/dl/src/test/scala/com/intel/analytics/bigdl/models/ModelGraientCheckSpec.scala
@@ -67,7 +67,7 @@ class ModelGraientCheckSpec extends FlatSpec with BeforeAndAfter with Matchers {
     val output = model.forward(input)
     val loss = criterion.forward(output, labels)
 
-    loss should be (6.9059443926654875)
+    loss should be (6.905944392665487)
   }
 
   "GoogleNet_v2 model in batch mode" should "be good in gradient check for input" in {