## What changes were proposed in this pull request?
Optimize LogSoftmax using MKL.

## How was this patch tested?
Unit Tests

For input = Tensor(20, 10000) and gradOutput = Tensor(20, 10000)

running on local desktop:
average elapsed time:

before = 0.029517 seconds
after = 8.99 * 1e-4 seconds

Boosting *32.8 times.
 

Test rnn model using spark-local with 4 cores
batchSize = 128
throughput is:
before:
174.9 records/s
after:
512.4 records/s



Tests on Broadwell with 44 cores.
one LSTM layer, throughput

outputSize before after
1000 	377.679 408.977
5000	346.314	331.818
10000	316.691	340.478
15000	306.378 338.387
20000	252.381 276.823